Add construct command to stembuild.
Write early version of construct command. Still have to flesh out what
it will actually do. As of now, we have some paramaters we think we will
need as well as a helper function to check and see if artifacts are
present in current-directory. Also add tests for our helper function.

[#161432730]

Co-authored-by: Malini Valliath <mvalliath@pivotal.io>